Force front end sort icons to not overlap the heading text
csvtotable sizes column width according to the widest string. If there is no content in the column wider than the header the text overlaps the sort buttons and makes them almost impossible to see. Trailing space along on the heading text did nothing. It was necessary to pad with spaces alternating with ESC control characters.