Only run unit tests before list generation
The point of doing the test before the list generation is to avoid generating the list from broken code. For those purposes, it doesn't matter if there's a typo in a comment or some trailing space in the readme and those should not block list generation (they are handled via the "on-push" workflow).