Give up on using Circle CI to generate list
The 5 hour maximum build duration of Circle CI is ending up to be too limiting to what I can do with the list. For now, I'll just manually generate the list. Hopefully in the future I'll find another automated generation solution since I'm likely to fail at keeping the list regularly updated.

Although it does have some very nice features, since I'm not using Circle CI for the list generation, I might as well go back to using Travis CI for continuous integration. The benefits are consistency with the rest of my repositories and the ability to use Travis Buddy to comment on PRs that fail CI. The unique features of Circle CI vs Travis CI are not useful for this particular application.