Filter Particle libraries from arduino library search
There's something called "Particle" (formerly "Spark") that is different from the Arduino IDE. It seems it is possible to make libraries compatible with both Arduino and Particle but my experience is the Particle people don't bother to do this but then they go and advertise the libraries as "Arduino". So I have to filter all that garbage out. Unfortunately this will cause false negatives for legitimate Arduino particle sensor libraries.

Unfortunately I can't filter out the "spark" keyword because that also takes out all the SparkFun libraries.