Change table heading padding character
It's necessary to use a non-printing character to force the web front end to not overwrite the sort button with the heading text when the heading text length exceeds the length of any content in the column. Previously, I was using ESC, but now that character shows up as squares in the browser, making it useless for this purpose. I'm now using a zero-width space instead.